requirements.txt
streamlit
pandas
plotly.express
matplotlib

